Hide qemu-bug until it is fixed.
Currently, we support only 32-bit guests anyway, so that it doesn't
hurt. Since I'm expecting it to take quite long until qemu fixes the bug
in the logical to linear address translation that occurs for 32-bit
guests on 64-bit hypervisors, we fix it this way, so that we can run
Escape as a guest.

The bug has been reported here:
https://bugs.launchpad.net/qemu/+bug/1211910